+++ HISTORY.rst
@@ -10,6 +10,8 @@ PENDING
 * Add ``mysqlparse.parse`` function.
 * Add ``six`` as a dependency to make writing Py2 and Py3 compatible code easier.
 * Add support for ``ALTER TABLE .. MODIFY [COLUMN]`` statements.
+* Add support for ``ALTER TABLE .. DROP *`` statements.
+
 
 0.1.3 (2016-02-20)
 ------------------